Ensure that the original secret isn't cleaned up if it was never replicated due to having no target prefix.

+++ controllers/replication/secret_controller.go
@@ -133,6 +133,10 @@ func (s *SecretSyncController) Reconcile(ctx context.Context, req ctrl.Request) 
 					}
 					for _, origSecret := range sourceSecretsToMapToTargets {
 						deleteObject := &corev1.Secret{ObjectMeta: metav1.ObjectMeta{Name: getPrefixedSecretName(target.TargetPrefix, origSecret.Name), Namespace: target.Namespace}}
+						if origSecret.Namespace == target.Namespace && origSecret.Name == deleteObject.Name {
+							// Target is the same secret as the original - bail.
+							continue
+						}
 						logger.Info("Deleting secrets for", "objectMeta", deleteObject.ObjectMeta,
 							"Cluster", target.K8sContextName)
 						err = remoteClient.Delete(ctx, deleteObject)
